mruby-regexp: fix integer overflow in large character classes
class_add_range doubled range_capa, a uint16_t, with no bound. A class
with 32769 or more non-ASCII codepoints grew it from 32768 to 65536,
which wrapped to 0; the following mrb_realloc(0) freed the buffer and
returned NULL, and the range was then written through the NULL pointer.
Widen num_ranges/range_capa to uint32_t and update range_capa only after
a successful realloc. Also merge a new range with the previous one when
they are contiguous or overlap, so an ascending run of codepoints (the
common case) collapses to a single range instead of one entry each.

Closes #6937.

+++ mrbgems/mruby-regexp/include/re_internal.h
@@ -51,8 +51,8 @@ typedef struct {
      ranges[2k] = lo, ranges[2k+1] = hi (inclusive). NULL when the
      class has no non-ASCII members (the common case). */
   uint32_t *ranges;
-  uint16_t num_ranges;
-  uint16_t range_capa;
+  uint32_t num_ranges;
+  uint32_t range_capa;
   mrb_bool negated;
   mrb_bool utf8_any;  /* match any non-ASCII byte if true */
 } re_charclass;
